appbackend/requirements.txt
pipenv
pymongo
datetime
python-dateutil
flask
flask-cors
gunicorn
certbot
certbot-nginx
sense-energy
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
oauthlib
requests-oauthlib
numpy
pandas
scikit-learn
joblib
